Remove eslint
It seems to create more problems than it solves, and most of the
things it keeps us safe from is already handled by typescript itself
and/or prettier.

Questions:

If there is some lints that actually adds value?  We could make a
typescpript-specific kth eslint configuration that fits better for our
use?

Does the other teams (still) use eslint?  Do they use it with typescript?